move namespace into Hashie::Mash
