+++ README.rdoc
@@ -10,14 +10,14 @@ You will now be able to access those keys through underscored key names (camelCa
 
 == Usage
 
-  @rash = Hashie::Rash.new({
-    "varOne" => 1, 
-    "two" => 2, 
-    :three => 3, 
-    :varFour => 4, 
-    "fiveHumpHumps" => 5, 
+  @rash = Hashie::Mash::Rash.new({
+    "varOne" => 1,
+    "two" => 2,
+    :three => 3,
+    :varFour => 4,
+    "fiveHumpHumps" => 5,
     :nested => {
-      "NestedOne" => "One", 
+      "NestedOne" => "One",
       :two => "two",
       "nested_three" => "three"
     },
@@ -26,7 +26,7 @@ You will now be able to access those keys through underscored key names (camelCa
       :nestedThree => 23
     }
   })
-  
+
   @rash.var_one                 # => 1
   @rash.two                     # => 2
   @rash.three                   # => 3
@@ -39,7 +39,7 @@ You will now be able to access those keys through underscored key names (camelCa
   @rash.nested_two.nested_three # => 23
 
 == Note on Patches/Pull Requests
- 
+
 * Fork the project.
 * Make your feature addition or bug fix.
 * Add tests for it. This is important so I don't break it in a
